requirements.txt
Django==4.2.5
PyJWT==2.8.0
django-soft-delete==0.9.21
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
psycopg2==2.9.9
python-dotenv==1.0.0
setuptools==68.2.2